Moved restricted location filter to common module
